requirements.txt
Django~=2.2.11
django-phonenumber-field~=5.0.0
django-model-utils~=4.0.0
phonenumbers
djangorestframework~=3.11.1
python-dateutil
requests
